# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DealerOwnerContext": {
        "properties": {
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "type": {
            "const": "dealer_organization",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "organization_id"
        ],
        "title": "DealerOwnerContext",
        "type": "object"
      },
...
      "ListingCreateRequest": {
        "properties": {
          "owner_context": {
            "discriminator": {
              "mapping": {
                "dealer_organization": "#/components/schemas/DealerOwnerContext",
                "personal": "#/components/schemas/PersonalOwnerContext"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/PersonalOwnerContext"
              },
              {
                "$ref": "#/components/schemas/DealerOwnerContext"
              }
            ],
            "title": "Owner Context"
          },
          "variant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Id"
          },
          "vehicle_type": {
            "enum": [
              "car",
              "bike"
            ],
            "title": "Vehicle Type",
            "type": "string"
          }
        },
        "required": [
          "owner_context",
          "vehicle_type"
        ],
        "title": "ListingCreateRequest",
        "type": "object"
      },
...
      "ListingPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ListingPrivateResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "ListingPageResponse",
        "type": "object"
      },
      "ListingPrivateResponse": {
        "properties": {
          "asking_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asking Price"
          },
          "bike_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BikeSpecInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "car_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CarSpecInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_user_id": {
            "format": "uuid",
            "title": "Created By User Id",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lifecycle_status": {
            "enum": [
              "draft",
              "removed"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "moderation_status": {
            "enum": [
              "not_started",
              "pending"
            ],
            "title": "Moderation Status",
            "type": "string"
          },
          "owner_organization_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Organization Id"
          },
          "owner_type": {
            "enum": [
              "user",
              "dealer_organization"
            ],
            "title": "Owner Type",
            "type": "string"
          },
          "owner_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner User Id"
          },
          "publication_status": {
            "enum": [
              "private",
              "pending"
            ],
            "title": "Publication Status",
            "type": "string"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "submitted_listing_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted Listing Version"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "variant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Id"
          },
          "vehicle_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VehicleSpecInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "vehicle_type": {
            "enum": [
              "car",
              "bike"
            ],
            "title": "Vehicle Type",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "owner_type",
          "owner_user_id",
          "owner_organization_id",
          "created_by_user_id",
          "vehicle_type",
          "variant_id",
          "lifecycle_status",
          "publication_status",
          "moderation_status",
          "submitted_listing_version",
          "submitted_at",
          "title",
          "description",
          "asking_price",
          "currency",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "ListingPrivateResponse",
        "type": "object"
      },
      "ListingSubmissionRequest": {
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "ListingSubmissionRequest",
        "type": "object"
      },
      "ListingUpdateRequest": {
        "properties": {
          "asking_price": {
            "anyOf": [
              {
                "maximum": 1000000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asking Price"
          },
          "bike_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BikeSpecInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "car_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CarSpecInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "vehicle_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VehicleSpecInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "ListingUpdateRequest",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "correlation_id"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "PublicationReadinessResponse": {
        "additionalProperties": false,
        "properties": {
          "evaluated_at": {
            "format": "date-time",
            "title": "Evaluated At",
            "type": "string"
          },
          "gates": {
            "items": {
              "$ref": "#/components/schemas/ReadinessGateResponse"
            },
            "title": "Gates",
            "type": "array"
          },
          "listing_id": {
            "format": "uuid",
            "title": "Listing Id",
            "type": "string"
          },
          "listing_version": {
            "title": "Listing Version",
            "type": "integer"
          },
          "moderation_status": {
            "enum": [
              "not_started",
              "pending"
            ],
            "title": "Moderation Status",
            "type": "string"
          },
          "ownership_reused": {
            "title": "Ownership Reused",
            "type": "boolean"
          },
          "publication_status": {
            "enum": [
              "private",
              "pending"
            ],
            "title": "Publication Status",
            "type": "string"
          },
          "publishable": {
            "const": false,
            "title": "Publishable",
            "type": "boolean"
          },
          "submission_attempt_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Attempt Id"
          },
          "submission_status": {
            "enum": [
              "not_submitted",
              "blocked",
              "verification_pending",
              "moderation_pending"
            ],
            "title": "Submission Status",
            "type": "string"
          }
        },
        "required": [
          "listing_id",
          "listing_version",
          "submission_attempt_id",
          "submission_status",
          "publication_status",
          "moderation_status",
          "ownership_reused",
          "publishable",
          "gates",
          "evaluated_at"
        ],
        "title": "PublicationReadinessResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "WheelMatch API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/listings": {
      "post": {
        "operationId": "create_listing_api_v1_listings_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingPrivateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Listing",
        "tags": [
          "listings"
        ]
      }
    },
    "/api/v1/listings/{listing_id}": {
      "get": {
        "operationId": "get_listing_api_v1_listings__listing_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingPrivateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Listing",
        "tags": [
          "listings"
        ]
      },
      "patch": {
        "operationId": "update_listing_api_v1_listings__listing_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingPrivateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Listing",
        "tags": [
          "listings"
        ]
      }
    },
...
    "/api/v1/listings/{listing_id}/publication-readiness": {
      "get": {
        "operationId": "get_publication_readiness_api_v1_listings__listing_id__publication_readiness_get",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Publication Readiness",
        "tags": [
          "listings"
        ]
      }
    },
    "/api/v1/listings/{listing_id}/submit": {
      "post": {
        "operationId": "submit_listing_api_v1_listings__listing_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Listing",
        "tags": [
          "listings"
        ]
      }
    },
...
    "/api/v1/me/listings": {
      "get": {
        "operationId": "list_owned_listings_api_v1_me_listings_get",
        "parameters": [
          {
            "in": "query",
            "name": "owner_type",
            "required": false,
            "schema": {
              "default": "personal",
              "pattern": "^(personal|dealer_organization)$",
              "title": "Owner Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "organization_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Organization Id"
            }
          },
          {
            "in": "query",
            "name": "lifecycle_status",
            "required": false,
            "schema": {
              "default": "draft",
              "pattern": "^(draft|removed)$",
              "title": "Lifecycle Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "minLength": 16,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Owned Listings",
        "tags": [
          "listings"
        ]
      }
    },
...
  }
}